NaLang/analyzer/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
spacy>=3.7.0
python-multipart>=0.0.9

# PDF support
pdfplumber>=0.11.0
pypdf>=4.0.0

# Other formats (optional)
python-docx>=1.1.0
beautifulsoup4>=4.12.0
striprtf>=0.0.26
